@@ -4,35 +4,44 @@
   "version": "4.2.1",
   "license": "Apache-2.0",
   "main": "./dist/src/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "require": "./dist/src/index.js",
+      "default": "./dist/src/index.js"
+    }
+  },
   "type": "module",
   "scripts": {
     "lint": "eslint .",
     "dev": "concurrently \"tsc --watch\" \"tsc-alias -w\"",
     "build": "yarn run lint && tsc --build --verbose --force && tsc-alias -p tsconfig.json",
     "clean": "rimraf dist",
-    "test": "yarn run lint && yarn run jest --coverage"
+    "test": "yarn run lint && NODE_OPTIONS='--experimental-require-module' yarn run jest --coverage"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.0",
-    "@babel/preset-env": "^7.24.0",
+    "@babel/core": "^7.26.9",
+    "@babel/preset-env": "^7.26.9",
+    "@babel/preset-typescript": "^7.26.0",
     "@jest/globals": "^29.7.0",
+    "@tsconfig/recommended": "^1.0.8",
     "@types/jest": "^29.5.14",
-    "@types/luxon": "^3",
+    "@types/luxon": "^3.4.2",
     "@typescript-eslint/eslint-plugin": "^8.24.1",
     "@typescript-eslint/parser": "^8.24.1",
     "babel-jest": "^29.7.0",
     "concurrently": "^9.1.2",
-    "eslint": "^9.20.1",
+    "eslint": "^9.21.0",
     "jest": "^29.7.0",
-    "rimraf": "^5.0.10",
+    "rimraf": "^6.0.1",
     "ts-jest": "^29.2.5",
     "tsc-alias": "^1.8.10",
     "typescript": "^5.7.3"
   },
   "dependencies": {
-    "@fjell/core": "^4.2.12",
-    "@fjell/http-api": "^4.2.0",
-    "@fjell/logging": "^4.2.11"
+    "@fjell/core": "^4.3.1",
+    "@fjell/http-api": "^4.3.0",
+    "@fjell/logging": "^4.3.4"
   },
   "packageManager": "yarn@4.6.0",
   "repository": {
