@@ -21,7 +21,7 @@ export const getAllActionOperation = <
     api: HttpApi,
     apiOptions: ClientApiOptions,
     utilities: Utilities<V, S, L1, L2, L3, L4, L5>
-    // eslint-disable-next-line max-params
+     
   ) => {
     
   const allAction = async (
