@@ -24,7 +24,7 @@ export const getOneOperation = <
     api: HttpApi,
     apiOptions: ClientApiOptions,
     utilities: Utilities<V, S, L1, L2, L3, L4, L5>
-    // eslint-disable-next-line max-params
+     
   ): (
     query: ItemQuery,
     options: Partial<GetMethodOptions>,
