@@ -23,7 +23,7 @@ export const getUpdateOperation = <
     api: HttpApi,
     apiOptions: ClientApiOptions,
     utilities: Utilities<V, S, L1, L2, L3, L4, L5>
-    // eslint-disable-next-line max-params
+     
   ) => {
     
   const update = async (
